requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# MySQL driver for TiDB
mysql-connector-python==8.2.0

# Optional: for better logging and development
python-multipart==0.0.6